+++ src/array/includes.js
@@ -1,7 +1,7 @@
 import curry from '../function/curry'
 
 /**
- * @name contains
+ * @name includes
  * @since v0.1.0
  * @category Array
  * @sig a → Array a → Boolean
@@ -12,15 +12,15 @@ import curry from '../function/curry'
  * @return {Boolean} A Boolean based on if the value is found or not
  *
  * @example
- * contains(3, [1, 2, 3]); // => true
+ * includes(3, [1, 2, 3]); // => true
  *
  * // It is also curried
  *
- * const checker = contains(3);
+ * const checker = includes(3);
  *
  * checker([1, 2, 3]); // => true
  */
-const contains = (a, list) =>
+const includes = (a, list) =>
   list.indexOf(a) !== -1
 
-export default curry(contains)
+export default curry(includes)